requirements.txt
numpy==1.23.5
pandas==1.5.3
psycopg2-binary==2.9.6
tushare==1.4.24
fastapi==0.119.1
pydantic==1.10.15
uvicorn==0.38.0
apscheduler==3.11.0
greenlet==3.2.4
akshare
requests==2.32.3
beautifulsoup4==4.12.3
py-mini-racer==0.10.0
pytest==8.4.2
